fix url and type of chainId block_num

+++ mystiko_indexer_client/src/types/commitment_included.rs
@@ -5,7 +5,7 @@ use typed_builder::TypedBuilder;
 #[serde(rename_all = "camelCase")]
 pub struct CommitmentIncludedFilter {
     #[builder(setter(strip_option), default=None)]
-    pub chain_id: Option<u32>,
+    pub chain_id: Option<u64>,
     #[builder(setter(strip_option), default=None)]
     pub contract_address: Option<String>,
     #[builder(setter(strip_option), default=None)]
@@ -13,7 +13,7 @@ pub struct CommitmentIncludedFilter {
     #[builder(setter(strip_option), default=None)]
     pub tx_hash: Option<String>,
     #[builder(setter(strip_option), default=None)]
-    pub block_num: Option<u32>,
+    pub block_num: Option<u64>,
     #[builder(setter(strip_option), default=None)]
     pub create_at: Option<u64>,
     #[builder(setter(strip_option), default=None)]
@@ -24,23 +24,23 @@ pub struct CommitmentIncludedFilter {
 
 #[derive(Deserialize, Serialize, TypedBuilder)]
 pub struct CommitmentIncludedForChainRequest {
-    pub chain_id: u32,
+    pub chain_id: u64,
     #[builder(setter(strip_option), default=None)]
-    pub start_block: Option<u32>,
+    pub start_block: Option<u64>,
     #[builder(setter(strip_option), default=None)]
-    pub end_block: Option<u32>,
+    pub end_block: Option<u64>,
     #[builder(setter(strip_option), default=None)]
     pub where_filter: Option<CommitmentIncludedFilter>,
 }
 
 #[derive(Deserialize, Serialize, TypedBuilder)]
 pub struct CommitmentIncludedForContractRequest {
-    pub chain_id: u32,
+    pub chain_id: u64,
     pub address: String,
     #[builder(setter(strip_option), default=None)]
-    pub start_block: Option<u32>,
+    pub start_block: Option<u64>,
     #[builder(setter(strip_option), default=None)]
-    pub end_block: Option<u32>,
+    pub end_block: Option<u64>,
     #[builder(setter(strip_option), default=None)]
     pub where_filter: Option<CommitmentIncludedFilter>,
 }
@@ -49,11 +49,11 @@ pub struct CommitmentIncludedForContractRequest {
 #[serde(rename_all = "camelCase")]
 pub struct CommitmentIncludedResponse {
     pub id: u64,
-    pub chain_id: u32,
+    pub chain_id: u64,
     pub contract_address: String,
     pub commit_hash: String,
     pub tx_hash: String,
-    pub block_num: u32,
+    pub block_num: u64,
     pub create_at: u64,
     pub status: u32,
     pub contract_id: u64,
